+++ FrontendReact/ale-beauty-react/src/components/SubCateProd.jsx
@@ -1,22 +1,16 @@
 import React, { useEffect, useState } from "react";
-import { useParams, useOutletContext, useNavigate } from "react-router-dom";
+import { useParams, useOutletContext } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import IconButton from "@mui/joy/IconButton";
-import FavoriteBorder from "@mui/icons-material/FavoriteBorder";
-import Favorite from "@mui/icons-material/Favorite";
-import Rating from "@mui/material/Rating";
 import CircularProgress from "@mui/material/CircularProgress";
 import Typography from "@mui/material/Typography";
 import "../assets/stylesheets/SubCateProds.css";
-import { formatCOP } from "../services/currency";
-import noImage from "../assets/images/no_image.png";
 import { useAlert } from "../components/AlertProvider.jsx";
+import ProductCard from "../components/ProductCard"; // ✅ IMPORTAR
 
 export default function ProductsPageSubCategory() {
   const { categorySlug, subCategorySlug, lang } = useParams();
   const { t } = useTranslation();
   const { favoriteIds, loadFavorites } = useOutletContext();
-  const navigate = useNavigate();
 
   const [token, setToken] = useState(null);
   const [products, setProducts] = useState([]);
@@ -35,32 +29,6 @@ export default function ProductsPageSubCategory() {
     if (savedToken) setToken(savedToken);
   }, []);
 
-  function GradientHeart({ filled = false, size = 36 }) {
-    return (
-      <svg
-        width={size}
-        height={size}
-        viewBox="0 0 24 24"
-        style={{ display: "block" }}
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <defs>
-          <linearGradient id="heart-gradient" x1="0" y1="0" x2="24" y2="24" gradientUnits="userSpaceOnUse">
-            <stop stopColor="#cf0d5bff"/>
-            <stop offset="1" stopColor="#f7c1dcff"/>
-          </linearGradient>
-        </defs>
-        <path
-          d="M12.1 18.55l-.1.1-.1-.1C7.14 14.24 4 11.39 4 8.5 4 6.5 5.5 5 7.5 5c1.54 0 3.04.99 3.57 2.36h1.87C13.46 5.99 14.96 5 16.5 5c2 0 3.5 1.5 3.5 3.5 0 2.89-3.14 5.74-7.9 10.05z"
-          fill={filled ? "url(#heart-gradient)" : "none"}
-          stroke="url(#heart-gradient)"
-          strokeWidth="2"
-        />
-      </svg>
-    );
-  }
-
   useEffect(() => {
     if (!token) return;
 
@@ -75,7 +43,6 @@ export default function ProductsPageSubCategory() {
         );
         if (!res.ok) throw new Error("Error al cargar los productos");
         const data = await res.json();
-        console.log("products response", data);
         const prods = data.products || [];
         setProducts(prods);
 
@@ -128,6 +95,7 @@ export default function ProductsPageSubCategory() {
   }, [token, categorySlug, subCategorySlug]);
 
   let filteredProducts = products;
+  
   if (priceRange.min || priceRange.max) {
     filteredProducts = filteredProducts.filter((p) => {
       const price = p.precio_producto;
@@ -137,12 +105,14 @@ export default function ProductsPageSubCategory() {
       );
     });
   }
+  
   if (selectedRatings.length > 0) {
     filteredProducts = filteredProducts.filter((p) => {
       const productRating = Math.floor(productRatings[p.id]?.avg || 0);
       return selectedRatings.includes(productRating);
     });
   }
+  
   if (sortOrder === "asc") {
     filteredProducts = [...filteredProducts].sort((a, b) => a.precio_producto - b.precio_producto);
   } else if (sortOrder === "desc") {
@@ -158,21 +128,22 @@ export default function ProductsPageSubCategory() {
         });
         if (res.ok) {
           await loadFavorites();
-          addAlert("se elimino de tus favoritos", "warning", 3500);
+          addAlert("Se eliminó de tus favoritos", "warning", 3500);
         }
       } else {
         const res = await fetch("https://localhost:4000/api/v1/favorites", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
-            Authorization: `Bearer ${token}` },
+            Authorization: `Bearer ${token}` 
+          },
           body: JSON.stringify({ product_id: productId }),
         });
         const data = await res.json();
         if (data.success) {
           await loadFavorites();
-          addAlert("se agregó a tus favoritos", "success", 3500);
-        };
+          addAlert("Se agregó a tus favoritos", "success", 3500);
+        }
       }
     } catch (err) {
       console.error("Error al cambiar favorito:", err);
@@ -181,6 +152,13 @@ export default function ProductsPageSubCategory() {
   };
 
   const addToCart = (productId) => {
+    // ✅ Optimistic update
+    window.dispatchEvent(new CustomEvent("cartUpdatedOptimistic", { 
+      bubbles: false,
+      detail: { productId, action: 'add' }
+    }));
+    addAlert("Se agregó al carrito", "success", 3500);
+
     fetch("https://localhost:4000/api/v1/cart/add_product", {
       method: "POST",
       headers: {
@@ -194,11 +172,21 @@ export default function ProductsPageSubCategory() {
         if (data.cart) {
           setCart(data.cart);
           window.dispatchEvent(new CustomEvent("cartUpdatedCustom", { bubbles: false }));
-          addAlert("se agregó al carrito", "success");
+        } else if (data.errors) {
+          addAlert(t('productDetails.error') + data.errors.join(", "), "error", 3500);
+          window.dispatchEvent(new CustomEvent("cartUpdateFailed", { 
+            bubbles: false,
+            detail: { productId, action: 'add' }
+          }));
         }
       })
       .catch((err) => {
         console.error(t("productDetails.cartAddError"), err);
+        addAlert(t("productDetails.cartAddError"), "error", 3500);
+        window.dispatchEvent(new CustomEvent("cartUpdateFailed", { 
+          bubbles: false,
+          detail: { productId, action: 'add' }
+        }));
       });
   };
 
@@ -237,23 +225,15 @@ export default function ProductsPageSubCategory() {
     <section className="shop-section">
       <div className="shop-wrapper">
         <div className="shop-header">
-        <div >
-        <h1
-        style={{
-            margin: 48,
-            fontSize: "3.5rem",
-            fontWeight: "bold",
-            color: "#000000ff",
-            fontFamily: "'Poppins', Arial, sans-serif",
-          }}
-        >
-          {subCategory ? toTitleCase(subCategory.nombre) : "PRODUCTOS"}
-        </h1>
+          <h1>
+            {subCategory ? toTitleCase(subCategory.nombre) : "PRODUCTOS"}
+          </h1>
         </div>
-        </div>
+        
         <div className="shop-container">
           <aside className="filter-sidebar">
             <h3>Filtros</h3>
+            
             <div className="filter-group">
               <h4>Ordenar por precio</h4>
               <label>
@@ -284,6 +264,7 @@ export default function ProductsPageSubCategory() {
                 Mayor a menor
               </label>
             </div>
+            
             <div className="filter-group">
               <h4>Rango de precio</h4>
               <div className="price-range">
@@ -306,6 +287,7 @@ export default function ProductsPageSubCategory() {
                 />
               </div>
             </div>
+            
             <div className="filter-group">
               <h4>Valoración</h4>
               {[4, 3, 2, 1].map((rating) => (
@@ -325,6 +307,7 @@ export default function ProductsPageSubCategory() {
                 </label>
               ))}
             </div>
+            
             {(sortOrder ||
               priceRange.min ||
               priceRange.max ||
@@ -341,58 +324,20 @@ export default function ProductsPageSubCategory() {
               </button>
             )}
           </aside>
+          
+          {/* ✅ GRID CON PRODUCTCARD */}
           <div className="products-grid">
-            {filteredProducts.map((prod) => (
-              <div className="custom-product-card product-card" key={prod.id} style={{ cursor: "pointer" }} onClick={() => navigate(`/${lang}/producto/${prod.slug}`)} tabIndex={0}
-                onKeyDown={e => { if (e.key === 'Enter' || e.key === ' ') { navigate(`/${lang}/producto/${prod.slug}`); } }}>
-                <div className="custom-image-wrapper">
-                  <img src={prod.imagen_url || noImage} alt={prod.nombre_producto} />
-                  <IconButton
-                    onClick={e => { e.stopPropagation(); toggleFavorite(prod.id); }}
-                    className="custom-favorite-btn"
-                    style={{ padding: 0, background: "transparent" }}
-                  >
-                    {favoriteIds.includes(prod.id)
-                      ? <GradientHeart filled />
-                      : <GradientHeart filled={false} />
-                    }
-                  </IconButton>
-                </div>
-                <div className="custom-product-info">
-                  <div className="custom-rating-row-v2">
-                    <svg width="17" height="17" viewBox="0 0 24 24" fill="#FFC107" style={{ marginRight: "2px", verticalAlign: "middle" }}>
-                      <path d="M12 17.27L18.18 21l-1.64-7.03L22 9.24l-7.19-.61L12 2 9.19 8.63 2 9.24l5.46 4.73L5.82 21z"/>
-                    </svg>
-                    <span className="custom-rating-number-v2">{productRatings[prod.id]?.avg ? Number(productRatings[prod.id]?.avg).toFixed(1) : "0.0"}</span>
-                  </div>
-                  <div className="custom-product-name-v2">{prod.nombre_producto}</div>
-                  <div className="custom-price-row-v2">
-                    {/* DESCUENTO */}
-                    {prod.precio_con_mejor_descuento && prod.precio_con_mejor_descuento < prod.precio_producto ? (
-                      <>
-                        <span className="custom-price-v2">{formatCOP(prod.precio_con_mejor_descuento)}</span>
-                        <span className="custom-price-v2 line-through">{formatCOP(prod.precio_producto)}</span>
-                      </>
-                    ) : (
-                      <span className="custom-price-v2">{formatCOP(prod.precio_producto)}</span>
-                    )}
-                  </div>
-                  {/* Nombre del descuento SOLO si existe y hay descuento */}
-                  {prod.mejor_descuento_para_precio && prod.precio_con_mejor_descuento < prod.precio_producto && (
-                    <div className="custom-descuento-nombre">
-                      {prod.mejor_descuento_para_precio.nombre}
-                      {prod.mejor_descuento_para_precio.tipo === "porcentaje"
-                        ? ` (${prod.mejor_descuento_para_precio.valor}%)`
-                        : ` (-${formatCOP(prod.mejor_descuento_para_precio.valor)})`}
-                    </div>
-                  )}
-                </div>
-                <div className="custom-card-footer">
-                  <button onClick={e => { e.stopPropagation(); addToCart(prod.id); }} className="custom-add-btn">
-                    {t("products.addToCart")}
-                  </button>
-                </div>
-              </div>
+            {filteredProducts.map((prod, index) => (
+              <ProductCard
+                key={prod.id}
+                product={prod}
+                lang={lang}
+                isFavorite={favoriteIds.includes(prod.id)}
+                onToggleFavorite={toggleFavorite}
+                onAddToCart={addToCart}
+                productRating={productRatings[prod.id]}
+                t={t}
+              />
             ))}
           </div>
         </div>
