+++ FrontendReact/ale-beauty-react/src/components/hooks/useCategories.js
@@ -1,0 +1,55 @@
+import { useState, useEffect, useCallback } from 'react';
+
+export const useCategories = () => {
+  const [categories, setCategories] = useState([]);
+  const [hoveredNavCategory, setHoveredNavCategory] = useState(null);
+  const [loading, setLoading] = useState(false);
+
+  const fetchCategories = useCallback(async () => {
+    setLoading(true);
+    try {
+      const token = localStorage.getItem('token');
+      const res = await fetch('https://localhost:4000/api/v1/categories', {
+        headers: token ? { Authorization: `Bearer ${token}` } : {},
+      });
+      
+      if (res.ok) {
+        const data = await res.json();
+        const categoriesArray = Array.isArray(data) ? 
+          data : 
+          Array.isArray(data.categories) ? 
+          data.categories : 
+          [];
+        
+        setCategories(categoriesArray);
+      }
+    } catch (error) {
+      console.error('Error fetching categories:', error);
+    } finally {
+      setLoading(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    fetchCategories();
+  }, [fetchCategories]);
+
+  // Precargar imágenes
+  useEffect(() => {
+    categories.forEach(cat => {
+      const img = new Image();
+      img.src = cat.imagen_url;
+      cat.sub_categories?.forEach(sub => {
+        const subImg = new Image();
+        subImg.src = sub.imagen_url;
+      });
+    });
+  }, [categories]);
+
+  return {
+    categories,
+    hoveredNavCategory,
+    setHoveredNavCategory,
+    loading
+  };
+};
