+++ FrontendReact/ale-beauty-react/src/components/hooks/useScrollState.js
@@ -1,0 +1,16 @@
+import { useState, useEffect } from 'react';
+
+export const useScrollState = (threshold = 100) => {
+  const [scrolled, setScrolled] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrolled(window.scrollY > threshold);
+    };
+
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, [threshold]);
+
+  return scrolled;
+};
