+++ FrontendReact/ale-beauty-react/src/pages/inicio/hooks/useCartActions.js
@@ -2,6 +2,16 @@ import { useCallback } from 'react';
 
 export const useCartActions = (token, setCart, addAlert, t) => {
   const addToCart = useCallback((productId) => {
+    // ✅ OPTIMISTIC UPDATE - Actualizar contador INMEDIATAMENTE
+    window.dispatchEvent(new CustomEvent("cartUpdatedOptimistic", { 
+      bubbles: false,
+      detail: { productId, action: 'add' }
+    }));
+    
+    // ✅ Mostrar alerta inmediatamente
+    addAlert("Se agregó al carrito", "success", 3500);
+
+    // Hacer la petición al servidor en background
     fetch("https://localhost:4000/api/v1/cart/add_product", {
       method: "POST",
       headers: {
@@ -13,6 +23,7 @@ export const useCartActions = (token, setCart, addAlert, t) => {
       .then((res) => res.json())
       .then((data) => {
         if (data.cart) {
+          // ✅ Actualizar con datos reales del servidor
           setCart(data.cart);
           window.dispatchEvent(new CustomEvent("cartUpdatedCustom", { bubbles: false }));
 
@@ -32,14 +43,24 @@ export const useCartActions = (token, setCart, addAlert, t) => {
             });
             console.log("🛒 Evento GA4 enviado: add_to_cart", product);
           }
-          addAlert("Se agregó al carrito", "success", 3500);
         } else if (data.errors) {
+          // ✅ Si falla, revertir y mostrar error
           addAlert(t('productDetails.error') + data.errors.join(", "), "error", 3500);
+          window.dispatchEvent(new CustomEvent("cartUpdateFailed", { 
+            bubbles: false,
+            detail: { productId, action: 'add' }
+          }));
         }
       })
       .catch((err) => {
         console.error(t('productDetails.cartAddError'), err);
         addAlert(t('productDetails.cartAddError'), "error", 3500);
+        
+        // ✅ Revertir cambio optimista
+        window.dispatchEvent(new CustomEvent("cartUpdateFailed", { 
+          bubbles: false,
+          detail: { productId, action: 'add' }
+        }));
       });
   }, [token, setCart, addAlert, t]);
